More unit testing / restructure
