--- tests/cached_nulecules/gitlab/answers.conf.sample
@@ -1,3 +1,0 @@
-[general]
-namespace = default
-provider = kubernetes
