+++ tests/units/cli/test_examples/gitlab/Dockerfile
