+++ tests/units/cli/test_examples/gitlab/artifacts/kubernetes/gitlab-rc.json
