+++ tests/units/cli/test_examples/gitlab/artifacts/kubernetes/postgres-rc.json
