+++ tests/units/cli/test_examples/gitlab/artifacts/kubernetes/postgres-service.json
