+++ tests/units/cli/test_examples/gitlab/artifacts/kubernetes/redis-service.json
