+++ tests/units/cli/test_examples/guestbook-go/Nulecule
