+++ tests/units/cli/test_examples/guestbook-go/artifacts/kubernetes/guestbook-controller.json
