+++ tests/units/cli/test_examples/helloapache/artifacts/docker/hello-apache-pod_run
