+++ tests/units/cli/test_examples/helloapache/artifacts/kubernetes/hello-apache-pod.json
