+++ tests/units/cli/test_examples/wordpress-centos7-atomicapp/artifacts/openshift/service.yaml
