+++ tests/units/cli/test_examples/wordpress-centos7-atomicapp/external/aggregated-mariadb-atomicapp/artifacts/kubernetes/mariadb-pod.yaml
