+++ tests/units/cli/test_examples/wordpress-centos7-atomicapp/external/aggregated-mariadb-atomicapp/gpl-3.0.txt
